requirements.txt
black
bs4
fastparquet
gradio
langchain
openai
pandas
pinecone-client
pyarrow
tiktoken
typer
unstructured
